requirements.txt
aiogram==3.18.0
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1